Shit happens

+++ backend/pkg/models/project.go
@@ -21,14 +21,14 @@ type Projects struct {
 }
 
 type ProjectAnalytic struct {
-	Id                 int     `json:"Id"`
-	Key                string  `json:"Key"`
-	Title              string  `json:"Name"`
-	AllIssuesCount     int     `json:"allIssuesCount"`
-	OpenIssuesCount    int     `json:"openIssuesCount"`
-	CloseIssuesCount   int     `json:"closeIssuesCount"`
-	AverageTime        float64 `json:"averageTime"`
-	AverageIssuesCount float64 `json:"averageIssuesCount"`
+	Id                 int    `json:"Id"`
+	Key                string `json:"Key"`
+	Title              string `json:"Name"`
+	AllIssuesCount     int    `json:"allIssuesCount"`
+	OpenIssuesCount    int    `json:"openIssuesCount"`
+	CloseIssuesCount   int    `json:"closeIssuesCount"`
+	AverageTime        int    `json:"averageTime"`
+	AverageIssuesCount int    `json:"averageIssuesCount"`
 }
 
 type PageCount struct {
